Another attempt. PublicDashboard and props from DashboardControls
