+++ frontend/src/metabase/dashboard/components/Dashboard/Dashboard.tsx
@@ -11,6 +11,7 @@ import type {
 } from "metabase/dashboard/actions";
 import { DashboardHeader } from "metabase/dashboard/components/DashboardHeader";
 import { DashboardControls } from "metabase/dashboard/hoc/DashboardControls";
+import type { DashboardControlsPassedProps } from "metabase/dashboard/hoc/types";
 import type {
   FetchDashboardResult,
   SuccessfulFetchDashboardResult,
@@ -69,8 +70,7 @@ import {
   DashboardEmptyStateWithoutAddPrompt,
 } from "./DashboardEmptyState/DashboardEmptyState";
 
-interface DashboardProps {
-  dashboardId: DashboardId;
+type DashboardProps = {
   route: Route;
   params: { slug: string };
   children?: ReactNode;
@@ -106,28 +106,10 @@ interface DashboardProps {
   isNavigatingBackToDashboard: boolean;
   addCardOnLoad?: DashCardId;
   editingOnLoad?: string | string[];
-  location: Location;
-  isNightMode: boolean;
-  isFullscreen: boolean;
-  hasNightModeToggle: boolean;
-  refreshPeriod: number | null;
 
   initialize: (opts?: { clearCache?: boolean }) => void;
-  fetchDashboard: (opts: {
-    dashId: DashboardId;
-    queryParams?: Record<string, unknown>;
-    options?: {
-      clearCache?: boolean;
-      preserveParameters?: boolean;
-    };
-  }) => Promise<FetchDashboardResult>;
-  fetchDashboardCardData: (opts?: {
-    reload?: boolean;
-    clearCache?: boolean;
-  }) => Promise<void>;
   fetchDashboardCardMetadata: () => Promise<void>;
   cancelFetchDashboardCardData: () => void;
-  loadDashboardParams: () => void;
   addCardToDashboard: (opts: {
     dashId: DashboardId;
     cardId: CardId;
@@ -138,7 +120,6 @@ interface DashboardProps {
   addLinkDashCardToDashboard: (opts: NewDashCardOpts) => void;
   archiveDashboard: (id: DashboardId) => Promise<void>;
 
-  onRefreshPeriodChange: (period: number | null) => void;
   setEditingDashboard: (dashboard: IDashboard | null) => void;
   setDashboardAttributes: (opts: SetDashboardAttributesOpts) => void;
   setSharing: (isSharing: boolean) => void;
@@ -189,15 +170,10 @@ interface DashboardProps {
     slug: string,
   ) => EmbeddingParameterVisibility | null;
   updateDashboardAndCards: () => void;
-  onFullscreenChange: (
-    isFullscreen: boolean,
-    browserFullscreen?: boolean,
-  ) => void;
 
-  onNightModeChange: () => void;
   setSidebar: (opts: { name: DashboardSidebarName }) => void;
   hideAddParameterPopover: () => void;
-}
+} & DashboardControlsPassedProps;
 
 function DashboardInner(props: DashboardProps) {
   const {
